attendanceLedger/requirements.txt
django == 2.0.7
mysql-connector-python == 8.0.12
mysqlclient == 1.3.13
djangorestframework == 3.8.2